requirements.txt
kafka-python 
requests 
boto3 
flask 
flask-cors
pyspark==3.5.1
findspark==2.0.1
mysql-connector-python